Moved cached transformer
