+++ sktime/transformers/segment.py
@@ -9,8 +9,8 @@ from sktime.utils.validation.supervised import validate_X
 from sktime.transformers.base import BaseTransformer
 from sktime.utils.data_container import tabularize, detabularize, concat_nested_arrays
 
-__author__ = ["Markus Löning", "Jason Lines", "Piotr Oleśkiewicz", "George Oastler"]
-__all__ = ['RandomIntervalSegmenter', 'IntervalSegmenter', 'DerivativeSlopeTransformer', 'TimeSeriesConcatenator', 'CachedTransformer']
+__author__ = ["Markus Löning", "Jason Lines", "Piotr Oleśkiewicz"]
+__all__ = ['RandomIntervalSegmenter', 'IntervalSegmenter', 'DerivativeSlopeTransformer', 'TimeSeriesConcatenator']
 
 class IntervalSegmenter(BaseTransformer):
     """
@@ -430,60 +430,3 @@ class PlateauFinder(BaseTransformer):
         Xt["%s_starts" % column_prefix] = pd.Series(self._starts)
         Xt["%s_lengths" % column_prefix] = pd.Series(self._lengths)
         return Xt
-
-
-class CachedTransformer(BaseTransformer):
-    """Transformer that transforms data and adds the transformed version to a cache. If the transformation is called again on already seen data the data is fetched from the cache rather than performing the expensive transformation.
-
-        Parameters
-        ----------
-        transformer : transformer
-            the transformer to transform uncached data
-        """
-
-    def __init__(self, transformer):
-        self.cache = {}
-        self.transformer = transformer
-
-    """
-        clear the cache
-    """
-    def clear(self):
-        self.cache = {}
-
-    def transform(self, X, y=None):
-        """
-        Fit transformer, creating a cache for transformation.
-
-        Parameters
-        ----------
-        X : pandas DataFrame of shape [n_samples, n_features]
-            Input data
-        y : pandas Series, shape (n_samples, ...), optional
-            Targets for supervised learning.
-
-        Returns
-        -------
-        self : an instance of self.
-        """
-        # for each instance, get transformed instance from cache or transform and add to cache
-        cached_instances = {}
-        uncached_indices = []
-        for index in X.index.values:
-            try:
-                cached_instances[index] = self.cache[index]
-            except:
-                uncached_indices.append(index)
-        if len(uncached_indices) > 0:
-            uncached_instances = X.loc[uncached_indices, :]
-            transformed_uncached_instances = \
-                self.transformer.transform(uncached_instances)
-            transformed_uncached_instances.index = uncached_instances.index
-            transformed_uncached_instances = transformed_uncached_instances.to_dict('index')
-            self.cache.update(transformed_uncached_instances)
-            cached_instances.update(transformed_uncached_instances)
-        cached_instances = pd.DataFrame.from_dict(cached_instances, orient='index')
-        return cached_instances
-
-    def __str__(self):
-        return self.transformer.__str__()
